Organize IPC latency test
