--- ipc_latency/c_tcp_wireshark.txt
@@ -1,11 +1,0 @@
-486119217 - 486039887,77054,178410
-486307382 - 486212333,93112,196423
-486503802 - 486410060,94342,187162
-486696629 - 486597385,98894,192589
-486882029 - 486790130,92743,185921
-487070076 - 486976379,94246,186849
-487260249 - 487163518,97499,189630
-487446779 - 487353319,94242,185856
-487650719 - 487539499,119619,168285
-487737203 - 487692487,43331,722210
-# Wireshark, One-Way, RoundTrip
