+++ ipc_latency/epoll/result.txt
@@ -1,0 +1,10 @@
+1718.770
+1755.479
+1724.905
+1764.713
+1695.104
+1730.109
+1710.175
+1748.521
+1710.843
+1706.795
