--- ipc_latency/go_tcp_wireshark.txt
@@ -1,11 +1,0 @@
-557744358 - 557704640,557738994 - 557699875,50998
-557782608 - 557767569,557778496 - 557765017,26868
-557870289 - 557804325,557866755 - 557798396,79163
-557893753 - 557884099,557891893 - 557881835,18191
-557923059 - 557916055,557921389 - 557913934,16171
-557957731 - 557950199,557956261 - 557944955,17609
-557981241 - 557972512,557979650 - 557968795,19089
-558000060 - 557994160,557998942 - 557992091,14068
-558022315 - 558014804,558021259 - 558012369,16189
-558041112 - 558034311,558039991 - 558032747,15329
-# Wireshark, One-Way, RoundTrip
