--- ipc_latency/result.txt
@@ -1,25 +1,0 @@
-Go:
-Wireshark       One-Way         RoundTrip
-39718           39119           50998
-15039           13479           26868
-65964           68359           79163
-9654            10058           18191
-7004            7455            16171
-7532            11306           17609
-8729            10855           19089
-5900            6851            14068
-7511            8890            16189
-6801            7244            15329
-
-C:
-Wireshark       One-Way       RoundTrip
-79330           77054         178410
-95049           93112         196423
-93742           94342         187162
-99244           98894         192589
-91899           92743         185921
-93697           94246         186849
-96731           97499         189630
-93460           94242         185856
-111220          119619        168285
-44716           43331         72221
