+++ ipc_latency/select/all.txt
@@ -1,0 +1,104 @@
+1:
+12.068
+12.506
+12.428
+12.133
+17.012
+12.606
+12.674
+12.279
+12.556
+12.519
+
+2:
+7.665
+7.801
+7.663
+7.582
+7.764
+7.799
+7.618
+7.614
+7.623
+7.634
+
+4:
+11.760
+11.817
+12.028
+12.529
+12.570
+11.982
+12.634
+12.216
+11.909
+12.577
+
+8:
+23.149
+24.087
+24.203
+23.650
+23.455
+23.891
+23.335
+24.021
+23.697
+23.742
+
+16:
+47.680
+41.904
+37.136
+46.872
+46.910
+47.361
+47.106
+42.617
+48.483
+43.019
+
+32:
+54.401
+53.262
+55.832
+56.574
+59.656
+51.899
+52.986
+54.302
+56.499
+53.333
+
+64:
+301.664
+193.708
+119.920
+297.496
+247.094
+193.199
+207.299
+155.953
+203.626
+173.276
+207.416
+294.611
+193.902
+237.508
+360.240
+171.334
+215.036
+258.918
+189.390
+249.360
+
+88.067
+101.004
+142.905
+90.284
+86.270
+88.524
+95.957
+120.014
+91.675
+79.972
