+++ ipc_latency/unix_domain_socket/clean.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+rm ./unix_domain_*
